requirements.txt
streamlit
pandas
matplotlib
fpdf2
google-generativeai
python-dotenv